feat(open-api): format generated client code and improve tagged operation type names
Format generated code through prettier so we don't need to ignore it in eslint settings.

Instead of deduplicating operation ids using a counter, we use an operations tags to prefix the
operation id, which results in more user-friendly names for generated request/response types.

While with the counter approach we could support duplicate operation ids without tags, we opt
instead to throw an error as it's not very intuitive to have numbered operations with the same name
in the client.