fix(trpc): share a single tanstack query client for all trpc apis
The useQueryClient hook finds the query client from the closest parent provider. This means that
when integrating with multiple tRPC APIs, we could only find the query client for the API defined
first. This change shares a single query client between all tRPC APIs, so that the useQueryClient
hook gives you the query client for any tRPC API.