feat(open-api): support for type-safe streaming responses in generated typescript client
This change adds support for OpenAPI operations to declare that they will return a streaming
response via setting the vendor extension 'x-streaming' to true. When set, the generated client will
provide an async iterator for consumers to iterate over each chunk as it becomes available, in a
type-safe manner.

re #19