feat(fast-api): improve operation names in open api specification for better generated clients
By default, FastAPI includes the function name, path and http method in the operationId which in
turn is used to generate the names of the functions to call each operation. For example:

```python
@app.get('/items')
def list_items(...):
	...
```

By default gets the operationId `list_items_items_get`, which isn't very friendly in the generated
client, which would end up having a method `client.listItemsItemsGet(...)`.

We use just the function name as the operationId to address this. Which for this example would be
`client.listItems(...)`