feat(api-connection): tanstack query hooks for connecting react to fastapi
This change implements the open-api#ts-hooks generator, which provides factory methods for options
to provide to tanstack query hooks, in a similar manner to tRPC.

This saves the need for users to manage state themselves, and provides parity with tRPC.

By default, PUT/PATCH/POST/DELETE http requests are considered "mutations" and all others "queries".
This behaviour can be overridden with `x-mutation` and `x-query` vendor extensions however.

Similarly to tRPC, if an operation accepts a `cursor` parameter, we consider this a paginated
operation and generate an `infiniteQueryOptions` method alongside the `queryOptions` method. If an
operation is paginated but has a different pagination parameter name, this can be specified using
the `x-cursor` vendor extension.

We support streaming operations in a similar manner to tRPC, where a streaming query hook will
continually have its data updated as chunks are received. Since a single cache entry is shared by
all pages for an infinite query hook, we cannot stream multiple pages at once for a "paginated
streaming" operation as this risks conflicts/race-conditions. We therefore iterate through the whole
stream for each page in this case.
For mutation operations which return a stream, we don't do anything special and just return the
iterator for the user to use if required.

Fixes #96