feat(api-connection): react -> fastapi share singleton vanilla client instance
The useXXXApiClient hook would construct a new client every time, which meant that credentials
cached by the useSigV4 hook wouldn't be reused across different pages, leading to a slight
performance hit for using the vanilla client instead of the hooks. This change updates the
XXXApiProvider component to also set up a context with the api client.