feat: add ts#nx-generator generator for generating generators (#140)
* feat: add ts#nx-generator generator for generating generators

Add a new generator which adds a generator to an existing typescript project.

When run within this repo, it adds some extra pieces specific to our generators
such as adding a guide to the docs.

Note that we remove the dist/xxx entry in the tsconfig.base.json paths since
this causes issues when running local generators, and doesn't seem to be
required for the build and/or IDE completion.

Fixes #127

* docs: update translations

* docs: fix schema import for nx generator guide

* docs: update translations

* fix(nx-generator): handle docs i18n, use bun 1.1 in ci

Update to ensure we write the guide in the 'en' folder now that docs are
internationalised.

Bun 1.2 uses a text-based lockfile as the default, which Nx does not
currently support, causing issues with @nx/dependency-checks in the e2e
tests for our ts#nx-generator generator.

* chore(nx-generator): docs and minor template changes

* docs: update translations

* docs: update readme

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>