feat: use type-safe per-operation lambda handlers for trpc and fastapi (#166)
* feat: add hidden open-api#ts-metadata generator in preparation for type-safe fastapi infrastructure

Add a generator which emits metadata about each operation in an OpenAPI specification (namely its
name, path and method), so that appropriate routes can be added to the CDK constructs to deploy the
API in a type-safe manner. Additionally ensure that the ts#nx-generator generator increments the
metric no matter the order of generators in generators.json

References #146

* feat: use type-safe per-operation lambda handlers for trpc and fastapi

This change updates both trpc and fastapi apis to use individual lambda functions for each operation
in the API, in order for customers to follow the principle of least privilege, granting the minimal
permissions for each lambda handler to perform its specific function. We achieve this in a type-safe
manner, which uses type inference for trpc and minimal openapi code generation for fastapi. We
provide options for deploying using either an API Gateway REST API or HTTP API.

Fixes #146

* docs: update translations

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>