feat(infra): provide deploy and deploy-ci targets
Specifying a path to a cloud assembly is desired for CI/CD pipelines as it removes the potential for
non-deterministic deployments. However, when developing locally it can cause confusion/friction when
users need to build prior to deploying in order for changes to their CDK code to take effect. We
update the "deploy" command to not specify the cloud assembly to improve the dev experience, but
also provide a "deploy-ci" target which should be used in pipelines.