fix(open-api): fix code generation issues with array requests and responses
There were two issues with array requests/responses. Firstly when the request body is an array, we'd
get a type error in the client typeof someArray === 'object' was always true. Secondly when the
response is an array of objects the options proxy would incorrectly import Array<ObjectType> which
is invalid syntax. Fixed both of these issues.

Fixes #201