fix(py): prefix python module names with scope to avoid conflicts with remote packages
When a python project was generated with the same name as a dependency (eg a FastAPI project being
named 'fastapi'), the build would fail due to uv using the local fastapi package conflicting and
dependencies not being installed. We address this by using the fully qualified name as the project
name, and setting the module name by default to <scope>_<name>. This can be overridden by providing
a "moduleName" to the generator.

Fixes #228